Move get_and_set_concurrently to providers/utils.py
