+++ ScoutSuite/providers/aws/facade/cloudtrail.py
@@ -1,6 +1,7 @@
 from ScoutSuite.providers.aws.facade.utils import AWSFacadeUtils
 from ScoutSuite.providers.utils import run_concurrently
 from ScoutSuite.providers.aws.facade.basefacade import AWSBaseFacade
+from ScoutSuite.providers.utils import get_and_set_concurrently
 
 
 class CloudTrailFacade(AWSBaseFacade):
@@ -8,7 +9,7 @@ class CloudTrailFacade(AWSBaseFacade):
         client = AWSFacadeUtils.get_client('cloudtrail', self.session, region)
         trails = await run_concurrently(
             lambda: client.describe_trails()['trailList'])
-        await AWSFacadeUtils.get_and_set_concurrently(
+        await get_and_set_concurrently(
             [self._get_and_set_status, self._get_and_set_selectors], trails, region=region)
 
         return trails
