+++ ScoutSuite/providers/aws/facade/elb.py
@@ -1,7 +1,7 @@
 from ScoutSuite.providers.aws.facade.utils import AWSFacadeUtils
 from ScoutSuite.providers.aws.facade.basefacade import AWSBaseFacade
 from ScoutSuite.providers.aws.utils import ec2_classic
-from ScoutSuite.providers.utils import run_concurrently
+from ScoutSuite.providers.utils import run_concurrently, get_and_set_concurrently
 
 import asyncio
 
@@ -27,7 +27,7 @@ class ELBFacade(AWSBaseFacade):
                 load_balancer['VpcId'] =\
                     load_balancer['VPCId'] if 'VPCId' in load_balancer and load_balancer['VPCId'] else ec2_classic
 
-            await AWSFacadeUtils.get_and_set_concurrently(
+            await get_and_set_concurrently(
                 [self._get_and_set_load_balancer_attributes], self.load_balancers_cache[region], region=region)
 
     async def _get_and_set_load_balancer_attributes(self, load_balancer: {}, region: str):
