+++ ScoutSuite/providers/aws/facade/elbv2.py
@@ -3,7 +3,7 @@ import asyncio
 from ScoutSuite.providers.aws.facade.utils import AWSFacadeUtils
 from ScoutSuite.providers.aws.facade.basefacade import AWSBaseFacade
 from ScoutSuite.providers.aws.utils import ec2_classic
-from ScoutSuite.providers.utils import run_concurrently
+from ScoutSuite.providers.utils import run_concurrently, get_and_set_concurrently
 
 
 class ELBv2Facade(AWSBaseFacade):
@@ -27,7 +27,7 @@ class ELBv2Facade(AWSBaseFacade):
                 load_balancer['VpcId'] =\
                     load_balancer['VpcId'] if 'VpcId' in load_balancer and load_balancer['VpcId'] else ec2_classic
 
-            await AWSFacadeUtils.get_and_set_concurrently(
+            await get_and_set_concurrently(
                 [self._get_and_set_load_balancer_attributes], self.load_balancers_cache[region], region=region)
 
     async def _get_and_set_load_balancer_attributes(self, load_balancer: dict, region: str):
