+++ ScoutSuite/providers/aws/facade/utils.py
@@ -78,28 +78,3 @@ class AWSFacadeUtils:
         return AWSFacadeUtils._clients.setdefault(
             (service, region),
             session.client(service, region_name=region) if region else session.client(service))
-
-    @staticmethod
-    async def get_and_set_concurrently(get_and_set_funcs: [], entities: [], **kwargs):
-        """
-        Given a list of get_and_set_* functions (ex: get_and_set_description, get_and_set_attributes,
-        get_and_set_policy, etc.) and a list of entities (ex: stacks, keys, load balancers, vpcs, etc.),
-        get_and_set_concurrently will call each of these functions concurrently on each entity.
-
-        :param get_and_set_funcs: list of functions that takes a region and an entity (they must have the following
-        signature: region: str, entity: {}) and then fetch and set some kind of attributes to this entity.
-        :param entities: list of a same kind of entities
-        :param region: a region
-
-        :return:
-        """
-
-        if len(entities) == 0:
-            return
-
-        tasks = {
-            asyncio.ensure_future(
-                get_and_set_func(entity, **kwargs)
-            ) for entity in entities for get_and_set_func in get_and_set_funcs
-        }
-        await asyncio.wait(tasks)
