+++ ScoutSuite/providers/utils.py
@@ -28,6 +28,31 @@ def run_concurrently(func):
     return asyncio.get_event_loop().run_in_executor(executor=None, func=func)
 
 
+async def get_and_set_concurrently(get_and_set_funcs: [], entities: [], **kwargs):
+    """
+    Given a list of get_and_set_* functions (ex: get_and_set_description, get_and_set_attributes,
+    get_and_set_policy, etc.) and a list of entities (ex: stacks, keys, load balancers, vpcs, etc.),
+    get_and_set_concurrently will call each of these functions concurrently on each entity.
+
+    :param get_and_set_funcs: list of functions that takes a region and an entity (they must have the following
+    signature: region: str, entity: {}) and then fetch and set some kind of attributes to this entity.
+    :param entities: list of a same kind of entities
+    :param region: a region
+
+    :return:
+    """
+
+    if len(entities) == 0:
+        return
+
+    tasks = {
+        asyncio.ensure_future(
+            get_and_set_func(entity, **kwargs)
+        ) for entity in entities for get_and_set_func in get_and_set_funcs
+    }
+    await asyncio.wait(tasks)
+
+
 async def map_concurrently(coro, entities, **kwargs):
     """
     Given a list of entities, executes coroutine `coro` concurrently on each entity and returns a list of the obtained
@@ -38,8 +63,11 @@ async def map_concurrently(coro, entities, **kwargs):
 
     :return: a list of new entities (ex: clusters)
     """
-    results = []
 
+    if len(entities) == 0:
+        return []
+
+    results = []
     tasks = {
         asyncio.ensure_future(
             coro(entity, **kwargs)
